requirements.txt
pandas
numpy
flask
mysql-connector-python
beautifulsoup4
scikit-learn
nltk
matplotlib
seaborn